Corrige tipos dos IDs nas classes de repositorio

+++ src/main/java/com/recrutaibackend/repository/InstitutionRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface InstitutionRepository extends JpaRepository<Institution, Long> {
+public interface InstitutionRepository extends JpaRepository<Institution, Integer> {
 }
